// graphql/gen/AttachmentBlockFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: AttachmentBlockFragment
// ====================================================

export interface AttachmentBlockFragment {
  __typename: "Attachment";
  id: number | null;
  title: string | null;
}